+++ apps/code-infra-dashboard/src/utils/formatters.ts
@@ -18,3 +18,42 @@ export function formatDiffMs(value: number): string {
   const sign = value > 0 ? '+' : '';
   return `${sign}${durationFormatter.format(value)} ms`;
 }
+
+interface ColumnDefinition {
+  field: string;
+  header?: string;
+  align?: 'left' | 'center' | 'right';
+}
+
+export function formatMarkdownTable(
+  columns: ColumnDefinition[],
+  data: Partial<Record<string, unknown>>[],
+): string {
+  let table = '';
+
+  const headers = columns.map((col) => col.header || col.field);
+  const alignments = columns.map((col) => col.align || 'left');
+
+  table += `| ${headers.join(' | ')} |\n`;
+
+  const separators = alignments.map((align) => {
+    switch (align) {
+      case 'center':
+        return ':---------:';
+      case 'right':
+        return '----------:';
+      case 'left':
+        return ':----------';
+      default:
+        return '-----------';
+    }
+  });
+  table += `|${separators.join('|')}|\n`;
+
+  data.forEach((row) => {
+    const cells = columns.map((col) => row[col.field] ?? '');
+    table += `| ${cells.join(' | ')} |\n`;
+  });
+
+  return table;
+}
